WICKET-7087: AjaxLazyLoadPanelTester not available in 10.0.0-M2
Introduce wicket-extensions-tester module that currently provides only
AjaxLazyLoadPanelTester

Add all **-tester modules to wicket-common-tests for better OSGi
compatibility.
