+++ wicket-extensions-tester/src/test/java/org/apache/wicket/extensions/tester/ajax/markup/html/AjaxLazyLoadPanelTest.java
@@ -14,8 +14,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.wicket.extensions.ajax.markup.html;
+package org.apache.wicket.extensions.tester.ajax.markup.html;
 
+import org.apache.wicket.extensions.ajax.markup.html.AjaxLazyLoadPanel;
+import org.apache.wicket.extensions.tester.AjaxLazyLoadPanelTester;
 import org.apache.wicket.util.tester.WicketTestCase;
 import org.junit.jupiter.api.Test;
